move logic classes into their own folder, use es6 goodness
